Convert security token in query parameter to cookie (#177)
This allows the correct token to be passed from a different location (e.g. zLUX iframe) to any endpoint in APIML as a cookie.

E.g.: `https://localhost:10010/ui/v1/apicatalog/?apimlAuthenticationToken=<correct_token>`

The request is then redirected to the URL without the parameter so it is not stored in the browser history.

In case of API catalog, it also redirects to the Dashboard because a request with empty path goes to log in even if the token is present.

Signed-off-by: Vasyl Melnychuk <vasyl.melnychuk@broadcom.com>