Fix of high cpu usage occurence replicated in Broadcom based on GIT issue #282 (#326)
Fix of high cpu usage occurrence replicated in Broadcom based on GIT issue #282
Solution is to switch LatencyUtils.useActualTime off, as described in javadoc of TimeServices
class. Moved gateway-common to apiml-common. Added check rule for method names
suppressed security module as there are some classes that triggered the check and this module is soon to be decommissioned.

Signed-off-by: janda06 <david.janda@broadcom.com>