Hardboiled/refactor/test speed (#506)
* Replace Tests testing framework with tests for the behavior of the application

The tests that were removed tested behavior of Java Enabler. These tests should
be and are in the library. Otherwise the tests are integration tests and should
be in integration tests.
The new tests verify only the application specific part.

* Improve configuration of what is old

The tests were waiting for few seconds. The time is a parameter of the tested
class. As such it is possible to speed up the tests.

* Remove test that does nothing.
For the test that valid https configuration will pass through provide mock.

Signed-off-by: Jakub Balhar <jakub.balhar@broadcom.com>