Add new logging for URL parameters
URLs can easily be malformed with poor configuration parameters. Relative path URLs are concatenated with the base URL, so the baseUrl parameter is expected not to have a trailing '/' and relative URLs are expected to begin with '/'. Logging has been added for the baseUrl, contextPath, homePageRelativeUrl, statusPageRelativeUrl, and healthCheckRelativeUrl parameters.

Signed-off-by: Carson Cook <carson.cook@ibm.com>