Add warning logging for poorly formed URL config parameters (#775)
* Fix intellij warnings in EurekaInstanceConfigValidator

Signed-off-by: Carson Cook <carson.cook@ibm.com>

* Fix intellij warnings in EurekaInstanceConfigValidatorTest

Signed-off-by: Carson Cook <carson.cook@ibm.com>

* Add new logging for URL parameters

URLs can easily be malformed with poor configuration parameters. Relative path URLs are concatenated with the base URL, so the baseUrl parameter is expected not to have a trailing '/' and relative URLs are expected to begin with '/'. Logging has been added for the baseUrl, contextPath, homePageRelativeUrl, statusPageRelativeUrl, and healthCheckRelativeUrl parameters.

Signed-off-by: Carson Cook <carson.cook@ibm.com>
Signed-off-by: Jakub Balhar <jakub.balhar@broadcom.com>